requirements.txt
python==2.7
django==1.8.2
djangorestframework==3.1.3
django-filter==0.10.0
django-mptt==0.7.4
django-honeypot==0.5.0
djangular==0.2.7
django-compressor==1.5
python-memcached==1.54
Pillow==2.9.0
lxml==3.4.4
pytz==2015.4
selenium==2.47.0
